requirements.txt
Twisted
storm
transaction
zope.component
zope.app.catalog
zope.app.intid
grokcore.component
grokcore.security
zope.securitypolicy
columnize
ipython>=0.11
ZODB3
pycrypto
pyOpenSSL
pyasn1
https://bitbucket.org/miracle2k/pyutils/get/pyutils.tar.gz
